package/ai/requirements.txt
# FastAPI & Server
fastapi==0.122.0
uvicorn[standard]==0.38.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# PaddleOCR & PaddlePaddle
paddleocr==3.3.2
paddlepaddle-gpu==3.2.0

# Computer Vision
opencv-python-headless>=4.9.0
Pillow>=10.0.0
numpy>=1.24.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
insightface>=0.7.3
onnxruntime-gpu>=1.15.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-json-logger>=2.0.7
psutil>=5.9.0
